# Conflicts:
#	libraries/applink/src/main/java/com/tokopedia/applink/internal/ApplinkConstInternalLogistic.kt

+++ features/transaction/purchase_platform_common/src/main/java/com/tokopedia/purchase_platform/common/feature/promo/view/model/validateuse/MvcShippingBenefitUiModel.kt
@@ -9,5 +9,5 @@ data class MvcShippingBenefitUiModel(
         @SerializedName("benefit_amount")
         var benefitAmount: Int = 0,
         @SerializedName("sp_id")
-        var spId: Int = 0
+        var spId: Long = 0
 ) : Parcelable
